% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{addIndex}
\alias{addIndex}
\title{Add index to repository}
\usage{
addIndex(service, catalogid = "root", repo = "", type, style = c(0, 1, 2))
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{type}{Type of index. Use listIndices(...,valid = TRUE) to see all possible indices}

\item{style}{Specifies the chunk style of the new index. Can be 0, 1 or 2}
}
\value{
An index of type to the triple-store. (invisible)
}
\description{
Add the index type from the triple-store.
}
\examples{
\dontrun{
service = createService('localhost','user','password')
addIndex(service,repo = "testRepo",type = "gspoi",style = 0)
}
}

