% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_functions.R
\name{addInitFile}
\alias{addInitFile}
\title{addInitFile}
\usage{
addInitFile(service, filepath, restart = "0")
}
\arguments{
\item{service}{Service object containing service url, username, and password}

\item{filepath}{The filepath to your initFile (text)}

\item{restart}{Boolean. If true, then any running shared backends will be restarted.}
}
\value{
ag put object: successful push or not
}
\description{
addInitFile
}
\examples{
\dontrun{
service = createService("localhost","user","password")
addInitFile(service,filepath = "path/to/directory/initFile.txt")
}
}

