% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_funcs.R
\name{addMappedType}
\alias{addMappedType}
\title{Add new mapped type}
\usage{
addMappedType(repository, type, encoding)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{type}{String}

\item{encoding}{String}
}
\description{
Add a new datatype mapping between type and encoding.
}

