% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{addNameSpace}
\alias{addNameSpace}
\title{addNameSpace}
\usage{
addNameSpace(service, catalogid = "root", repo = "newTest", prefix, nsURI)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{prefix}{Prefix}

\item{nsURI}{Namespace URI}
}
\value{
ag put object that states whether or not push was successful
}
\description{
addNameSpace
}
\examples{
\dontrun{
service = createService("localhost","user","password")
addNameSpace(service,catalogid = "root",repo = "testRepo",prefix = "tmp",
             nsURI = "http://test.com/tmp#")
}
}

