% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_funcs.R
\name{addPredicateMapping}
\alias{addPredicateMapping}
\title{Add new predicate mapping}
\usage{
addPredicateMapping(repository, predicate, encoding)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{predicate}{String}

\item{encoding}{String}
}
\description{
Add a new predicate mapping between predicate and encoding.
}

