% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{addStatement}
\alias{addStatement}
\title{addStatement}
\usage{
addStatement(service, catalogid = "root", repo = "testRepo", subj = NULL,
  pred = NULL, obj = NULL, context = NULL, session = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{subj}{valid url}

\item{pred}{valid url}

\item{obj}{valid url}

\item{context}{context of triple}

\item{session}{Your session object. If specified, this function will only make changes within the session.
You'll need to use sessionCommit() to integrate your additions.}
}
\value{
Return: successful push or not
}
\description{
Add a triple to the repository.
The subj, pred, and obj arguments must be supplied and are used to fill in the subject, predicate and object parts of the triple. The context argument is used to fill in the triple's graph but may be left off. In this case, the triple's graph will be the default graph of the repository.

The subj, pred, obj, and context (if supplied) arguments must all be in URI encoded N-Triples notation.

The service returns the triple-ID of the newly added triple.
}
\examples{
\dontrun{
service = createService('localhost','user','password')
subj = '<www.test.com/tmp#person>'
pred = '<www.test.com/tmp#hasItem>'
obj= '<www.test.com/tmp#sword>'
addStatement(service,catalogid = 'root',reposirepo = 'testRepo',
subj = subj,pred = pred,obj = obj)
}
}

