% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{addStatements}
\alias{addStatements}
\title{Add a list of triples}
\usage{
addStatements(repository, quads)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{quads}{A list of character vectors, where each vector representes a triples and has 4 elements.
The graph component can be NULL or ""}
}
\description{
Add a list of triples
}
\examples{
\dontrun{
service = createService("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
triples = list(c("<a>","<b>","<c>","<g1>"),
               c("<a>","<b>","<c>","<g2>"))
addStatements(rep,triples)
}
}

