% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{addStatementsFromFile}
\alias{addStatementsFromFile}
\title{Add statements from a file}
\usage{
addStatementsFromFile(repository, filepath, baseURI = NULL, context = NULL,
  commitEvery = NULL)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{filepath}{File that contains your triples}

\item{baseURI}{...}

\item{context}{Context to attach to added triples}

\item{commitEvery}{...}
}
\value{
an ag object that says whether or not the push was successful.
}
\description{
To be added
}
\examples{
\dontrun{
service = createService("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
addStatementsFromFile(rep,
file = 'path/to/file/mytriples.nq')
}
}

