% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allegRo.R
\docType{package}
\name{allegRo}
\alias{allegRo}
\alias{allegRo-package}
\title{allegRo: A package for dealing with RDF data and AllegroGraph}
\description{
This package interacts with AllegroGraph through a REST interface,
and provides many of the common actions one may want to execute in AllegroGraph.
In addition to this package, Python, Java, Lisp and Ruby clients all exist, but working
with AG from R allows the user to seamlessly integrate advanced analytics into their triplestore
}

