% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{backupRepository}
\alias{backupRepository}
\title{Create a backup for the  repository}
\usage{
backupRepository(service, catalogid = "root", repo = "", backupName)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{backupName}{character value, name of backup repository}
}
\description{
Create a backup of the repository.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
backupRepository(service,catalogid = "root",repo = "testRepo",backupName = "backup_testRepo")
}
}

