% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_funcs.R
\name{Bulkmode}
\alias{Bulkmode}
\alias{disableBulkMode}
\alias{enableBulkMode}
\alias{getBulkMode}
\title{getBulkMode}
\usage{
getBulkMode(repository)

enableBulkMode(repository)

disableBulkMode(repository)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}
}
\description{
getBulkMode
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
getBulkMode(rep)
enableenableBulkMode(rep)
disableBdisableBulkMode(rep)
}
}

