% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{createRepository}
\alias{createRepository}
\title{createRepository}
\usage{
createRepository(service, catalogid = "root", repo = "testFromR2",
  expectedSize = NULL, index = NULL, override = c("true", "false",
  "if-not-open"), restore = NULL, nocommit = c("1", "0"))
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{expectedSize}{Specifies the expected size of the repository.}

\item{index}{Can be specified mulitple times. Should hold index names, and is used to configure the set of indices created for the store.}

\item{override}{Override repository, 1 or 0.}

\item{restore}{Restore the repository.}

\item{nocommit}{...}
}
\value{
An object of ag with the response and given url.
}
\description{
createRepository
}
\examples{
\dontrun{
service = createService("localhost","user","password")
createRepository(service,catalogid = "root",repo = "test",
expectedSize = 100,index = NULL,override = "true",restore = NULL,nocommit = 1)
}
}

