% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_funcs.R
\name{deleteAllMappings}
\alias{deleteAllMappings}
\title{Delete all mappings in repository}
\usage{
deleteAllMappings(repository)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}
}
\description{
Remove all datatype and predicate mappings (included the default ones).
}

