% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{deleteAllNameSpaces}
\alias{deleteAllNameSpaces}
\title{Delete all namespaces in repository}
\usage{
deleteAllNameSpaces(repository, reset = FALSE)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{reset}{Defaults to FALSE. If TRUE,then the namespaces are reset to the default namespaces of the server.}
}
\value{
An ag delete object that states whether delete was successful or not (invisible)
}
\description{
Deletes all namespaces in this repository for the current user.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
deleteAllNameSpaces(rep,reset = TRUE)
}
}

