% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{deleteDuplicates}
\alias{deleteDuplicates}
\title{Delete all duplicate statements}
\usage{
deleteDuplicates(repository, mode = NULL, commit = NULL)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{mode}{A string (default: spog)}

\item{commit}{Commit defaults to true if the connection has autocommit mode on.}
}
\value{
An ag delete object that states whether delete was successful or not (invisible)
}
\description{
Remove duplicate triples from the repository
}
\examples{
\dontrun{
service = createService("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
deleteDuplicates(rep)
}
}

