% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{deleteFreeTextIndex}
\alias{deleteFreeTextIndex}
\title{Delete a free-text index}
\usage{
deleteFreeTextIndex(repository, indexName)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{indexName}{Name of index to delete}
}
\value{
TRUE or FALSE depending on success of deletion
}
\description{
Return a list of freetext indices defined on the store
}
\examples{
\dontrun{
service = createService("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
deleteFreeTextIndex(rep,repo = 'testRepo')
}
}

