% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_funcs.R
\name{deleteMappedType}
\alias{deleteMappedType}
\title{Delete a mapped type}
\usage{
deleteMappedType(repository, type)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{type}{String}
}
\description{
Remove the datatype mapping between type and encoding.
}

