% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_funcs.R
\name{deleteMappings}
\alias{deleteMappings}
\title{Delete user-defined mappings in repository}
\usage{
deleteMappings(repository)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}
}
\description{
Remove any user-defined datatype and predicate mappings.
}

