% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{deleteNameSpace}
\alias{deleteNameSpace}
\title{Delete a namespace in repository}
\usage{
deleteNameSpace(service, catalogid = "root", repo = "", namespace)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{namespace}{Namespace id to delete}
}
\value{
An ag delete object that states whether delete was successful or not (invisible)
}
\description{
Deletes a namespace in this repository for the current user.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
deleteNameSpace(service,catalogid = "root",repo = "testRepo")
}
}

