% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_funcs.R
\name{disableBulkMode}
\alias{disableBulkMode}
\title{disableBulkMode}
\usage{
disableBulkMode(service, catalogid = "root", repo = "")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}
}
\value{
ag put object
}
\description{
disableBulkMode
}
\examples{
\dontrun{
service = createService("localhost","user","password")
disableBulkMode(service,catalogid = "root",repo = 'testRepo')
}
}

