% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{dropIndex}
\alias{dropIndex}
\title{Drop an index from the repository}
\usage{
dropIndex(repository, type)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{type}{Type of index. Use listIndices(...,valid = TRUE) to see all possible indices}
}
\value{
TRUE or FALSE stating whether or not deletion was successful (invisible)
}
\description{
Drop the index type from the triple-store.
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
dropIndex(rep,type = "gspoi",style = 0)
}
}

