% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{evalQuery}
\alias{evalQuery}
\title{evalQuery}
\usage{
evalQuery(repository, query, returnType = c("data.table", "dataframe",
  "matrix", "list"), infer = NULL, context = NULL, cleanUp = FALSE,
  namedContext = NULL, bindings = NULL, planner = NULL,
  checkVariables = NULL, limit = NULL, convert = FALSE)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{query}{Sparql query to evaluate}

\item{returnType}{In what format should the triples be returned. Can choose between "dataframe","matrix","list". Defaults to "list}

\item{infer}{...}

\item{context}{Can be a list. Specifies the contexts from which you want to query}

\item{cleanUp}{If TRUE, removes the XML types of all variables. Currently for this to work, the return type needs to be a data.table. However, this is recommended either way.}

\item{namedContext}{...}

\item{bindings}{...}

\item{planner}{...}

\item{checkVariables}{...}

\item{limit}{How many triples to return}

\item{convert}{Convert variable to R type}
}
\value{
Triples in matrix form
}
\description{
Execute a SPARQL select query
}
\examples{
\dontrun{
query = "select ?s ?p ?o {?s ?p ?o}"
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"root")
evalQuery(rep,
query = query, returnType = "data.table",
cleanUp = TRUE, limit = 10)
}
}

