% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service functions.R
\name{findProtocol}
\alias{findProtocol}
\title{List all catalogs}
\usage{
findProtocol(service, catalogid = "root")
}
\arguments{
\item{service}{Service object containing service url, username, and password}

\item{catalogid}{id for catalog of interest}
}
\value{
ag object
}
\description{
List all catalogs
}
\examples{
findProtocol(service)
}

