% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_funcs.R
\name{geospatial}
\alias{geospatial}
\alias{getCartesianGeoType}
\alias{getSphericalGeoType}
\alias{getStatementInsideCircle}
\alias{getStatementsHaversine}
\alias{getStatementsInsideBox}
\alias{listGeoTypes}
\title{Geospatial Datatype designer}
\usage{
listGeoTypes(repository)

getCartesianGeoType(repository, stripWidth = NULL, xMin = NULL,
  xMax = NULL, yMin = NULL, yMax = NULL)

getSphericalGeoType(repository, stripWidth = NULL, unit = "degree",
  latMin = NULL, latMax = NULL, longMin = NULL, longMax = NULL)

getStatementsInsideBox(repository, type, predicate, xMin, xMax, yMin, yMax,
  limit = NULL)

getStatementInsideCircle(repository, type, predicate, x, y, radius,
  limit = NULL, offset = NULL)

getStatementsHaversine(repository, type, predicate, lat, long, radius,
  radUnit = "km", limit = NULL, offset = NULL)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{xMin}{Helps determine size of cartesian plane}

\item{xMax}{Helps determine size of cartesian plane}

\item{yMin}{Helps determine size of cartesian plane}

\item{yMax}{Helps determine size of cartesian plane}

\item{unit}{Defaults to 'degree'. The unit in which stripWidth is specified. It can be degree, km, radian or mile.}

\item{latMin}{Limit the size of the region modelled by appropriate type}

\item{latMax}{Limit the size of the region modelled by appropriate type}

\item{type}{Specifies a geospatial sub-type, as created by getCartesianGeoType or getSphericalGeoType}

\item{predicate}{Specifies a predicate for the query}

\item{limit}{Specifies a maximum number of triples to return}

\item{radius}{Specifies the radius of the circle}

\item{offset}{Specifies how many triples should be skipped over}

\item{lat}{Specifies the latitude of the center of the circle}

\item{long}{Specifies the longitude of the center of the circle}

\item{radUnit}{Defaults to km. Spefies the unit in which radius is given when using Haversine}

\item{stripwidth}{Determines the granularity of the strip.}

\item{latMin}{Limit the size of the region modelled by appropriate type}

\item{latMin}{Limit the size of the region modelled by appropriate type}

\item{useContext}{When true (default false), use the context (graph) field of the triple instead of its object}
}
\description{
See official AllegroGraph documentation for information on 2D and nD Geospatial database design.
See examples below to get a general idea on how to use these functions.
}
\examples{

}

