% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{geoSpatialLiterals}
\alias{createCartesianGeoLiteral}
\alias{createSphericalGeoLiteral}
\alias{geoSpatialLiterals}
\title{Create a geospatial literal}
\usage{
createCartesianGeoLiteral(type, x, y)

createSphericalGeoLiteral(type, lat, long, unit = "degree")
}
\arguments{
\item{type}{Type created with getCartesianGeoType() or getSphericalGeoType()}

\item{x}{X coordinate}

\item{y}{Y coordinate}

\item{lat}{Latitude}

\item{long}{longitude}

\item{unit}{Defaults to 'degree'}
}
\description{
Helper function to create a cartesian literal using the type created with getCartesianGeoType or getSphericalGeoType
}
\keyword{literals,geospatial}

