% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_funcs.R
\name{getBulkMode}
\alias{getBulkMode}
\title{getBulkMode}
\usage{
getBulkMode(service, catalogid = "root", repo = "")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}
}
\value{
True if bulk mode is on, False otherwise
}
\description{
getBulkMode
}
\examples{
\dontrun{
service = createService("localhost","user","password")
getBulkMode(service,catalogid = "root",repo = 'testRepo')
}
}

