% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{getContexts}
\alias{getContexts}
\title{getContexts}
\usage{
getContexts(service, catalogid = "root", repo = "testRepo")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}
}
\value{
An ag get objects that displays all context id's
}
\description{
getContexts
}
\examples{
\dontrun{
service = createService("localhost","user","password")
getSize(service,catalogid = "root",repo = "testRepo")
}
}

