% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_funcs.R
\name{getInitFile}
\alias{getInitFile}
\title{getInitFile}
\usage{
getInitFile(service)
}
\arguments{
\item{service}{Service object containing service url, username, and password}
}
\value{
ag get object containing init file
}
\description{
getInitFile
}
\examples{
\dontrun{
service = createService("localhost","user","password")
getInitFile(service)
}
}

