% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{getSize}
\alias{getSize}
\title{getSize}
\usage{
getSize(service, catalogid = "root", repo = "", session = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{session}{If working in a session, specify response of createSession call}
}
\value{
An ag get object that shows how many triples are in the repository
}
\description{
getSize
}
\examples{
\dontrun{
service = createService("localhost","user","password")
getSize(service,catalogid = "root",repo = "testRepo")
}
}

