% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_funcs.R
\name{getVersion}
\alias{getVersion}
\title{getVersion}
\usage{
getVersion(service, complex = TRUE)
}
\arguments{
\item{service}{Service object containing service url, username, and password}

\item{complex}{Default TRUE. Will return all version related info. If FALSE, will just return version number}
}
\value{
Version of allegroGraph
}
\description{
getVersion
}
\examples{
\dontrun{
service = createService("localhost","user","password")
getVersion(service)
}
}

