% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_funcs.R
\name{listActiveSessions}
\alias{listActiveSessions}
\title{Retrieve information about active sessions}
\usage{
listActiveSessions(service)
}
\arguments{
\item{service}{Service object containing service url, username, and password}
}
\description{
Retrieve information about the currently running sessions.
}
\examples{
\dontrun{
service = service("localhost","user","password")
listActiveSessions
}
}

