% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_funcs.R
\name{listCatalogs}
\alias{listCatalogs}
\title{List all catalogs}
\usage{
listCatalogs(service)
}
\arguments{
\item{service}{Service object containing server url, username, and password}
}
\value{
For each catalog, id and uri properties are returned, giving respectively the name of the catalog and the URL under which it is found.
Properties named readable and writable indicate, for each catalog, whether the current user has read or write access to it.
}
\description{
Returns a set of catalogs that are available on this server.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
listCatalogs(service)
}
}

