% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{listContexts}
\alias{listContexts}
\title{List all contexts in repository}
\usage{
listContexts(repository)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}
}
\value{
list of all contexts
}
\description{
List all contexts in repository
}
\examples{
\dontrun{
service = createService("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
listContexs(rep)
}
}

