% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{listFreeTextIndices}
\alias{listFreeTextIndices}
\title{List free-text indices}
\usage{
listFreeTextIndices(repository)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}
}
\value{
A matrix of all active free text indices
}
\description{
Return a list of freetext indices defined on the store
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
listFreeTextIndices(rep)
}
}

