% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{listIndices}
\alias{listIndices}
\title{Return indices in repository}
\usage{
listIndices(repository, valid = FALSE)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{valid}{Default FALSE. If true, then all possible index flavors are returned rather than just the ones that are currently defined.}
}
\value{
List of active indices, or all possible valid indices if 'valid' is set to TRUE
}
\description{
Return a list of indices defined on the triple-store
}

