% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_funcs.R
\name{listMappings}
\alias{listMappings}
\title{List all mappings in repository}
\usage{
listMappings(repository)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}
}
\value{
A matrix containing the datatype and predicate mappings for the repository.
}
\description{
Return the datatype and predicate mappings for the repository.
}

