% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{listNameSpaces}
\alias{listNameSpaces}
\title{Return all namespaces in repository}
\usage{
listNameSpaces(repository)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}
}
\value{
An ag object that includes the list of namespaces used in the specified repository.
}
\description{
Return the currently active namespaces for your user in this repository
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
listNameSpaces(rep)
}
}

