% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{listRepositories}
\alias{listRepositories}
\title{List all repositories in chosen catalog}
\usage{
listRepositories(catalog, all = FALSE)
}
\arguments{
\item{catalog}{Catalog object specifying username, password, and catalog to work on.}

\item{all}{Default FALSE. If TRUE, then all repositories of all catalogs will be returned}
}
\value{
A list of all repositories in the catalog. Will show all repositories if all is set to TRUE
}
\description{
Return a list of repositories available in the catalog.
}
\examples{
\dontrun{
service = service("localhost","user","password")
cat = catalog(service,"root")
listRepositories(cat)
}
}

