% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{listScripts}
\alias{listScripts}
\title{getInitFile}
\usage{
listScripts(service, catalogid = "root", repositoryid = "")
}
\arguments{
\item{service}{Service object containing service url, username, and password}
}
\value{
ag get object containing init file
}
\description{
getInitFile
}
\examples{
\dontrun{
service = createService("localhost","user","password")
getInitFile(service)
}
}

