% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_class_functions.R
\name{print.ag_get}
\alias{print.ag_get}
\title{Print}
\usage{
\method{print}{ag_get}(x, ...)
}
\arguments{
\item{x}{The response from functions that receive info from the server}

\item{...}{}
}
\value{
The first 10 results from the request
}
\description{
Print
}
\examples{
\dontrun{
print(service)
}
}

