% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createService.R
\name{service}
\alias{service}
\title{Create a service object}
\usage{
service(url, user = NULL, password = NULL, testConnection = FALSE)
}
\arguments{
\item{url}{The url for your AllegroGraph server}

\item{user}{Username for user accessing store, if necessary}

\item{password}{Password for user accessing store, if necessary}

\item{testConnection}{Should the connection be tested? It can be tested seperately as well. Defaults to FALSE}
}
\value{
S3 Object of type "service", which states the url, username, and password for the user
}
\description{
Create a service object
}
\examples{
createService("localhost","user","password",testConnection = FALSE)
}

