% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_funcs.R
\name{setInitFile}
\alias{setInitFile}
\title{Set the initialization file}
\usage{
setInitFile(service, filepath, restart = FALSE)
}
\arguments{
\item{service}{Service object containing service url, username, and password}

\item{filepath}{The filepath to your initFile (text)}

\item{restart}{Defaults to FALSE. If TRUE, then any running shared backends will be restarted. This ensures that subsequent requests will use the new initialization file.}
}
\value{
ag put object stating the success of the push, or not (invisible)
}
\description{
Replace the current initialization file
}
\examples{
\dontrun{
service = service("localhost","user","password")
setInitFile(service,filepath = "path/to/directory/initFile.txt")
}
}

