% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_funcs.R
\name{startSession}
\alias{startSession}
\title{startSession}
\usage{
startSession(service, catalogid = "root", repo = "testFromR2",
  autocommit = NULL, lifetime = 100, loadInitFile = NULL, script = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{autocommit}{Bolean. Specify whether or not the session should use transactions.}

\item{lifetime}{The number of seconds the session can be idle before being shutdown and reclaimed.}

\item{loadInitFile}{If true, then the initfile will be loaded when the session starts.}

\item{script}{The name of a script file to load; may be specified multiple times.}
}
\value{
ag_post object containing the url of the session on the server
}
\description{
startSession
}
\examples{
\dontrun{
service = createService("localhost","user","password")
startSession(service,repo = "testRepo",lifetime = 1000,script = "path/to/script")
}
}

