% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createService.R
\name{summary.service}
\alias{summary.service}
\title{Summary.service}
\usage{
\method{summary}{service}(x, ...)
}
\arguments{
\item{x}{The service object}

\item{...}{}
}
\value{
Printout statement of your service parameters
}
\description{
Summary.service
}
\examples{
\dontrun{
service = createService("localhost","user","password")
summary(service)
}
}

