% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createService.R
\name{testConnection.service}
\alias{testConnection.service}
\title{testConnection}
\usage{
\method{testConnection}{service}(x, ...)
}
\arguments{
\item{x}{Your service object}

\item{...}{}
}
\value{
Statement that states whether or not you can proceed with service object
}
\description{
testConnection
}
\examples{
service = createService("localhost","user","password")
testConnection(service)
}

