% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{uploadMatrix}
\alias{uploadMatrix}
\title{Upload a data.table from a file}
\usage{
uploadMatrix(rep, triples, context = NULL)
}
\arguments{
\item{rep}{Repository object}

\item{triples}{data.table containing triples as three columns.}

\item{context}{graph name}
}
\description{
Upload a matrix in nquads format (matrix that is n x 3 or n x 4). Is it a wrapper around the
addStatementFile that efficiently writes out the triples to a temp file using data.table's fwrite, and
imports it into the repository
}

