% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funcs.R
\name{user}
\alias{addUser}
\alias{changeUserPassword}
\alias{deleteUser}
\alias{listUsers}
\alias{user}
\title{user}
\usage{
listUsers(service)

addUser(service, name, password)

deleteUser(service, name)

changeUserPassword(service, name, password)
}
\arguments{
\item{service}{The service object specifying port,username, and password}

\item{name}{Username}

\item{password}{Password associated with username}
}
\description{
Collection of functions to operate on different users on the server
}
\examples{
\dontrun{
service = service("localhost","user","password")
listUsers(service)
}
}

