% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{addIndex}
\alias{addIndex}
\title{Add index to repository}
\usage{
addIndex(repository, type, style = c(0, 1, 2))
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{type}{Type of index. Use listIndices(...,valid = TRUE) to see all possible indices}

\item{style}{Specifies the chunk style of the new index. Can be 0, 1 or 2}
}
\value{
An index of type to the triple-store. (invisible)
}
\description{
Add the index type from the triple-store.
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
addIndex(rep,type = "gspoi",style = 0)
}
}

