% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_functions.R
\name{addNameSpace}
\alias{addNameSpace}
\title{addNameSpace}
\usage{
addNameSpace(service, catalogid = "root", repositoryid = "newTest", prefix,
  nsURI)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repositoryid}{Id for repository of interest.}

\item{prefix}{Prefix}

\item{nsURI}{Namespace URI}
}
\value{
ag put object
}
\description{
addNameSpace
}

