% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{addStatement}
\alias{addStatement}
\title{addStatement}
\usage{
addStatement(service, catalogid = "root", repositoryid = "testRepo",
  subj = "<s>", pred = "<o>", obj = "<p>", context = NULL,
  session = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repositoryid}{Id for repository of interest.}

\item{subj}{valid url}

\item{pred}{valid url}

\item{obj}{valid url}

\item{context}{context of triple}

\item{session}{Your session object. If specified, this function will only make changes within the session.
You'll need to use sessionCommit() to integrate your additions.}
}
\value{
Return: successful push or not
}
\description{
addStatement
}
\examples{
\dontrun{
service = createService("localhost","user","password")
subj = "<www.test.com/tmp#person>"
pred = "<www.test.com/tmp#hasItem>"
obj= "<www.test.com/tmp#sword>"
addStatement(service,catalogid = "root",reposirepositoryid = "testRepo",
subj = subj,pred = pred,obj = obj)
}
}

