% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{addStatementsFromFile}
\alias{addStatementsFromFile}
\title{Add statements from a file}
\usage{
addStatementsFromFile(service, catalogid = "root", repo = "", filepath,
  baseURI = NULL, context = NULL, commitEvery = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{filepath}{File that contains your triples}

\item{baseURI}{...}

\item{context}{Context to attach to added triples}

\item{commitEvery}{...}
}
\value{
an ag object that says whether or not the push was successful.
}
\description{
To be added
}
\examples{
\dontrun{
service = createService("localhost","user","password")
addStatementsFromFile(service,catalogid = "root", repo = "testRepo",
file = "path/to/file/mytriples.nq")
}
}

