% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{addStatementsFromFile}
\alias{addStatementsFromFile}
\title{addStatementsFromFile}
\usage{
addStatementsFromFile(service, catalogid = "root", repositoryid = "",
  filepath, baseURI = NULL, context = NULL, commitEvery = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repositoryid}{Id for repository of interest.}

\item{filepath}{File that contains your triples}

\item{baseURI}{...}

\item{context}{...}

\item{commitEvery}{...}
}
\value{
an ag object that says whether or not the push was successful.
}
\description{
addStatementsFromFile
}
\examples{
\dontrun{
service = createService("localhost","user","password")
addStatementsFromFile(service,catalogid = "root", repositoryid = "testRepo",
file = "path/to/file/mytriples.nq")
}
}

