% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{backupRepository}
\alias{backupRepository}
\title{Create a backup for the  repository}
\usage{
backupRepository(repository, backupName)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{backupName}{character value, name of backup repository}
}
\value{
AllegroGraph default reponse
}
\description{
Create a backup of the repository.
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
backupRepository(,backupName = "backup_testRepo")
}
}

