% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{catalog}
\alias{catalog}
\title{Create a catalog object}
\usage{
catalog(service, catalog)
}
\arguments{
\item{service}{Object of type service on which your catalog is located}

\item{catalog}{Character string specifying your catalog}
}
\value{
S3 Object of type "catalog", which states the url, username, and password and catalog
}
\description{
Create a catalog object
}
\examples{
\dontrun{
service = service("localhost","user","password",testConnection = FALSE)
cat = catalog(service,"root")
}
}

