% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{createRepository}
\alias{createRepository}
\title{Create a repository}
\usage{
createRepository(catalog, repo, expectedSize = NULL, index = NULL,
  override = TRUE, restore = NULL, nocommit = NULL)
}
\arguments{
\item{catalog}{Catalog object specifying username, password, and catalog to work on.}

\item{repo}{Id for repository you want to create.}

\item{expectedSize}{Specifies the expected size of the repository.}

\item{index}{Can be specified mulitple times in a list. Should hold index names, and is used to configure the set of indices created for the store.}

\item{override}{Default TRUE}

\item{restore}{Restore the repository.}

\item{nocommit}{Boolean}
}
\value{
An object of ag with the response and given url.
}
\description{
Creates a new, empty repository.
When a repository with the given name already exists, it is overwritten, unless a parameter override with value FALSE is passed.
}
\examples{
\dontrun{
service = service("localhost","user","password")
cat = catalog(service,"root")
createRepository(cat,repo = "test",
expectedSize = 100,index = NULL,override = "true",restore = NULL,nocommit = 1)
}
}

