% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{deleteAllNameSpaces}
\alias{deleteAllNameSpaces}
\title{Delete all namespaces in repository}
\usage{
deleteAllNameSpaces(service, catalogid = "root", repo = "", reset = FALSE)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{reset}{Defaults to FALSE. If TRUE,then the namespaces are reset to the default namespaces of the server.}
}
\value{
An ag delete object that states whether delete was successful or not (invisible)
}
\description{
Deletes all namespaces in this repository for the current user.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
deleteAllNameSpaces(service,catalogid = "root",repo = "testRepo")
}
}

