% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{deleteDuplicates}
\alias{deleteDuplicates}
\title{deleteDuplicates}
\usage{
deleteDuplicates(service, catalogid = "root", repo = "testRepo",
  mode = NULL, commit = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{mode}{A string (default: spog)}

\item{commit}{A string}
}
\value{
An ag delete object
}
\description{
deleteDuplicates
}
\examples{
\dontrun{
service = createService("localhost","user","password")
deleteDuplicates(service,repo = "testRepo")
}
}

