% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{deleteNameSpace}
\alias{deleteNameSpace}
\title{Delete a namespace in repository}
\usage{
deleteNameSpace(repository, namespace)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{namespace}{Namespace id to delete}
}
\value{
An ag delete object that states whether delete was successful or not (invisible)
}
\description{
Deletes a namespace in this repository for the current user.
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
deleteNameSpace(service,catalogid = "root",repo = "testRepo",namespace = "foaf")
}
}

