% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_functions.R
\name{deleteNameSpaces}
\alias{deleteNameSpaces}
\title{deleteNameSpaces}
\usage{
deleteNameSpaces(service, catalogid = "root",
  repositoryid = "testWithParsa", reset = c(0, 1, "0", "1"))
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repositoryid}{Id for repository of interest.}

\item{reset}{Boolean, If 1, then the namespaces are reset to the default namespaces of the server.}
}
\value{
An ag delete object that states whether delete was successful or not
}
\description{
deleteNameSpaces
}
\examples{
\dontrun{
service = createService("localhost","user","password")
deleteNameSpaces(service,catalogid = "root",repositoryid = "testRepo")
}
}

