% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{deleteRepository}
\alias{deleteRepository}
\title{Delete a Repository}
\usage{
deleteRepository(catalog, repo)
}
\arguments{
\item{catalog}{Catalog object specifying username, password, and catalog to work on.}

\item{repo}{Id for repository you want to create.}
}
\value{
Logical statement in which TRUE implies a successful deletion (invisible)
}
\description{
Delete the specified repository. This can fail if the repository is being accessed by another client.
}
\examples{
\dontrun{
service = service("localhost","user","password")
cat = catalog(service,"root")
createRepository(cat,repo = "repotodelete")
deleteRepository(cat,repo = "repotodelete")
}
}

