% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{deleteRepository}
\alias{deleteRepository}
\title{Delete a Repository}
\usage{
deleteRepository(service, catalogid = "root", repo = "")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}
}
\value{
Logical statement in which TRUE implies a successful deletion (invisible)
}
\description{
Delete the specified repository. This can fail if the repository is being accessed by another client.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
createRepository(service,catalogid = "root",repo = "repotodelete")
deleteRepository(service,catalogid = "root",repo = "repotodelete")
}
}

