% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{deleteStatements}
\alias{deleteStatements}
\title{deleteStatements}
\usage{
deleteStatements(service, catalogid = "root", repositoryid = "testRepo",
  subj = NULL, pred = NULL, obj = NULL, context = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repositoryid}{Id for repository of interest.}

\item{subj}{valid url}

\item{pred}{valid url}

\item{obj}{valid url}

\item{context}{context of triple}
}
\value{
Return: successful push or not
}
\description{
deleteStatements
}
\examples{
\dontrun{
service = createService("localhost","user","password")
subj = "<www.test.com/tmp#person>"
pred = "<www.test.com/tmp#hasItem>"
obj= "<www.test.com/tmp#sword>"
deleteStatements(service,catalogid = "root",reposirepositoryid = "testRepo",
subj = subj,pred = pred,obj = obj)
}
}

