% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{deleteStatements}
\alias{deleteStatements}
\title{Delete matching statements}
\usage{
deleteStatements(service, catalogid = "root", repo = "testRepo",
  subj = NULL, pred = NULL, obj = NULL, context = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{subj}{valid url}

\item{pred}{valid url}

\item{obj}{valid url}

\item{context}{context of triple}
}
\value{
Return: successful push or not
}
\description{
Delete statements (triples) by matching against their components.
All parameters are optional -- when none are given, every statement in the store is deleted.
The patterns to match come in pairs for subject, predicate, object, and graph. When given, they should be specified in N-triples notation.
}
\examples{
\dontrun{
service = createService('localhost','user','password')
subj = '<www.test.com/tmp#person>'
pred = '<www.test.com/tmp#hasItem>'
obj= '<www.test.com/tmp#sword>'
deleteStatements(service,catalogid = 'root',reposirepo = 'testRepo',
subj = subj,pred = pred,obj = obj)
}
}

