% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{dropIndex}
\alias{dropIndex}
\title{Drop an index from the repository}
\usage{
dropIndex(service, catalogid = "root", repo = "", type)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{type}{Type of index. Use listIndices(...,valid = TRUE) to see all possible indices}
}
\value{
TRUE or FALSE stating whether or not deletion was successful (invisible)
}
\description{
Drop the index type from the triple-store.
}
\examples{
\dontrun{
service = createService('localhost','user','password')
dropIndex(service,repo = "testRepo",type = "gspoi",style = 0)
}
}

