% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freetextIndices_funcs.R
\name{evalFreeTextSearch}
\alias{evalFreeTextSearch}
\title{Query using free-text indices}
\usage{
evalFreeTextSearch(service, catalogid = "root", repo = "", pattern,
  index = NULL, limit = NULL, offset = NULL, sorted = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{pattern}{A string to match triples on}

\item{index}{The index you want to use}

\item{limit}{Integer, number of triples to return}

\item{offset}{Integer, number of triples to skip}

\item{sorted}{TRUE or FALSE, defaults to FALSE}
}
\value{
A list describing all configuration parametes of index of interest
}
\description{
Query the repository using a free-text index
}
\examples{
\dontrun{
service = createService('localhost','user','password')
createFreeTextIndex(service,repo = 'testRepo',index = 'index')
getFreeTextIndexConfiguration(service,catalogid = 'root',repo = 'testRepo',index = 'index')
}
}

