% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{evalFreeTextSearch}
\alias{evalFreeTextSearch}
\title{Query using free-text indices}
\usage{
evalFreeTextSearch(repository, pattern, index = NULL, limit = NULL,
  offset = NULL, sorted = NULL)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{pattern}{A string to match triples on}

\item{index}{The index you want to use}

\item{limit}{Integer, number of triples to return}

\item{offset}{Integer, number of triples to skip}

\item{sorted}{TRUE or FALSE, defaults to FALSE}
}
\value{
A list describing all configuration parametes of index of interest
}
\description{
Query the repository using a free-text index
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
getFreeTextIndexConfiguration(rep,repo = 'testRepo',index = 'index')
}
}

