% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freetextIndices_funcs.R
\name{evalFreeTextSearch}
\alias{evalFreeTextSearch}
\title{evalFreeTextSearch}
\usage{
evalFreeTextSearch(service, catalogid = "root", repo = "", pattern,
  index = NULL, limit = NULL, offset = NULL, sorted = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{pattern}{A string to match triples on}

\item{index}{The index you want to see configurations for}

\item{limit}{Number of triples to return}

\item{offset}{Number of triples to skip}

\item{sorted}{Boolean, defaults to FALSE}
}
\value{
A list describing all configuration parametes of index of interest
}
\description{
evalFreeTextSearch
}
\examples{
\dontrun{
service = createService("localhost","user","password")
createFreeTextIndex(service,repo = "testRepo",index = 'index')
getFreeTextIndexConfiguration(service,catalogid = "root",repo = 'testRepo',index = "index")
}
}

