% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{getAccess}
\alias{getAccess}
\title{getAccess}
\usage{
getAccess(service, catalogid = "root", repositoryid = "testRepo")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repositoryid}{Id for repository of interest.}
}
\value{
An ag get object shows the kind of access the user has on this repository
}
\description{
getAccess
}
\examples{
\dontrun{
service = createService("localhost","user","password")
getAccess(service,catalogid = "root",repositoryid = "testRepo")
}
}

