% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{getAccess}
\alias{getAccess}
\title{Repository access}
\usage{
getAccess(repository)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}
}
\value{
String describing access of user.
}
\description{
Return the access permissions the current user has for this repository.
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
getAccess(rep)
}
}

