% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{getAccess}
\alias{getAccess}
\title{Repository access}
\usage{
getAccess(service, catalogid = "root", repo = "testRepo")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}
}
\value{
String describing access of user.
}
\description{
Return the access permissions the current user has for this repository.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
getAccess(service,catalogid = "root",repo = "testRepo")
}
}

