% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{getDuplicates}
\alias{getDuplicates}
\title{Get all duplicate statements}
\usage{
getDuplicates(repository)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}
}
\value{
A matrix containing all duplicate statements
}
\description{
Get all duplicate statements
}
\section{Warning}{

This is slow on large repositories, even if no duplicates exist.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
getDuplicates(rep)
}
}

