% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{getDuplicates}
\alias{getDuplicates}
\title{getDuplicates}
\usage{
getDuplicates(service, catalogid = "root", repo = "testRepo")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}
}
\value{
A matrix containing all duplicate statements
}
\description{
getDuplicates
}
\section{Warning}{

This is slow on large repositories, even if no duplicates exits.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
print(getDuplicates(service,repo = "testRepo"))
}
}

