% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freetextIndices_funcs.R
\name{getFreeTextIndexConfiguration}
\alias{getFreeTextIndexConfiguration}
\title{getFreeTextIndexConfiguration}
\usage{
getFreeTextIndexConfiguration(service, catalogid = "root", repo = "", index)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{index}{The index you want to see configurations for}
}
\value{
A list describing all configuration parametes of index of interest
}
\description{
getFreeTextIndexConfiguration
}
\examples{
\dontrun{
service = createService("localhost","user","password")
createFreeTextIndex(service,repo = "testRepo",index = 'index')
getFreeTextIndexConfiguration(service,catalogid = "root",repo = 'testRepo',index = "index")
}
}

