% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{getFreeTextIndexConfiguration}
\alias{getFreeTextIndexConfiguration}
\title{Return free-text index configuration}
\usage{
getFreeTextIndexConfiguration(repository, index)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{index}{The index you want to see configurations for}
}
\value{
A list describing all configuration parametes of index of interest
}
\description{
Return information on the freetext-index named indexName.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
createFreeTextIndex(service,repo = 'testRepo',index = 'index')
getFreeTextIndexConfiguration(rep,repo = 'testRepo',index = 'index')
}
}

