% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_funcs.R
\name{getInitFile}
\alias{getInitFile}
\title{Find the contents of the Initialization file}
\usage{
getInitFile(service)
}
\arguments{
\item{service}{Service object containing service url, username, and password}
}
\value{
The text contained in your Initialization file
}
\description{
Retrieve the server's initialization file.
The initialization-file is a collection of Common Lisp code that is executed on every shared back-end when it starts
}
\examples{
\dontrun{
service = service("localhost","user","password")
getInitFile(service)
}
}

