% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{getSize}
\alias{getSize}
\title{Return the size of a repository}
\usage{
getSize(repository, context = NULL)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{context}{Can be a list. If supplied, the count is only for the specified named graphs}
}
\value{
Number of triples in store
}
\description{
Return the size of a repository
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
getSize(rep)
}
}

