% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{getStatements}
\alias{getStatements}
\title{addStatement}
\usage{
getStatements(service, catalogid = "root", repositoryid = "testRepo",
  subj = "s", pred = "o", obj = "p", context = NULL, infer = "false",
  callback = NULL, limit = NULL, tripleIDs = "false", count = "false")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repositoryid}{Id for repository of interest.}

\item{subj}{valid url}

\item{pred}{valid url}

\item{obj}{valid url}

\item{context}{context of triple}
}
\value{
Return: successful push or not
}
\description{
addStatement
}
\examples{
\dontrun{
service = createService("localhost","user","password")
subj = "<www.test.com/tmp#person>"
pred = "<www.test.com/tmp#hasItem>"
obj= "<www.test.com/tmp#sword>"
}
}

