% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{getStatements}
\alias{getStatements}
\title{Get statements}
\usage{
getStatements(repository, subj = NULL, pred = NULL, obj = NULL,
  context = NULL, infer = FALSE, limit = NULL, tripleIDs = FALSE,
  count = FALSE, returnType = c("data.table", "dataframe", "matrix",
  "list"), cleanUp = FALSE, convert = FALSE)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}

\item{subj}{Subject triple pattern you want to match - optional}

\item{pred}{Predicate of triple pattern you want to match - optional}

\item{obj}{Object of triple pattern you want to match - optional}

\item{context}{Context of triple pattern you want to match - optional}

\item{infer}{Specifies the kind of inference to use for this query. Defaults to FALSE, can also be 'rdfs++' or 'hasvalue'}

\item{limit}{How many triples should be returned}

\item{tripleIDs}{...}

\item{count}{...}

\item{returnType}{What type the return value should be. Best performance achieved with matrix.}

\item{cleanUp}{Should the xml schemas be removed. Has to be true for converting to R types to work}

\item{convert}{Convert variable to appropriate R type}
}
\value{
Triples based on the specified pattern
}
\description{
Can specify any part of statement, and all matches will be returned.
If no subject, predicate, or object supplied, all statements will be returned.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
getStatements(rep,
pred = '<http://test.org/pred>', limit = 10)
}
}

