% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement_funcs.R
\name{getStatements}
\alias{getStatements}
\title{getStatements}
\usage{
getStatements(service, catalogid = "root", repo = "testRepo", subj = NULL,
  pred = NULL, obj = NULL, context = NULL, infer = FALSE,
  limit = NULL, tripleIDs = FALSE, count = FALSE,
  returnType = c("data.table", "dataframe", "matrix", "list"),
  cleanUp = FALSE, convert = FALSE)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{subj}{Subject triple pattern you want to match - optional}

\item{pred}{Predicate of triple pattern you want to match - optional}

\item{obj}{Object of triple pattern you want to match - optional}

\item{context}{Context of triple pattern you want to match - optional}

\item{infer}{Specifies the kind of inference to use for this query. Defaults to FALSE, can also be "rdfs++" or "hasvalue"}

\item{limit}{How many triples should be returned}

\item{tripleIDs}{...}

\item{count}{...}

\item{returnType}{What type the return value should be. Best performance achieved with matrix.}

\item{cleanUp}{Should the xml schemas be removed. Has to be true for converting to R types to work}

\item{convert}{Convert variable to appropriate R type}
}
\value{
Triples based on the specified pattern
}
\description{
Can specify any part of statement, and all matches will be returned.
If no subject, predicate, or object supplied, all statements will be returned.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
getStatements(service,catalogid = "root",repo = "testRepo",
pred = "<http://test.org/pred>", limit = 10)
}
}

