% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_funcs.R
\name{getVersion}
\alias{getVersion}
\title{getVersion}
\usage{
getVersion(service)
}
\arguments{
\item{service}{}
}
\value{
Version of allegroGraph
}
\description{
getVersion
}
\examples{
\dontrun{
service = createService("localhost","user","password")
getVersion(service)
}
}

