% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service functions.R
\name{listCatalogs}
\alias{listCatalogs}
\title{List all catalogs}
\usage{
listCatalogs(service)
}
\arguments{
\item{service}{Service object containing service url, username, and password}
}
\value{
ag object
}
\description{
List all catalogs
}
\examples{
listCatalogs(service)
}

