% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{listContexts}
\alias{listContexts}
\title{List all contexts in repository}
\usage{
listContexts(service, catalogid = "root", repo = "")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}
}
\value{
list of all contexts
}
\description{
List all contexts in repository
}
\examples{
\dontrun{
service = createService("localhost","user","password")
listContexs(service,catalogid = "root",repo = "testRepo")
}
}

