% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freetextIndices_funcs.R
\name{listFreeTextIndices}
\alias{listFreeTextIndices}
\title{listFreeTextIndices}
\usage{
listFreeTextIndices(service, catalogid = "root", repo = "")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}
}
\value{
A matrix of all active free text indices
}
\description{
listFreeTextIndices
}
\examples{
\dontrun{
service = createService("localhost","user","password")
listFreeTextIndices(service,catalogid = "root",repo = 'testRepo')
}
}

