% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{listIndices}
\alias{listIndices}
\title{Return indices in repository}
\usage{
listIndices(service, catalogid = "root", repo = "", valid = FALSE)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{valid}{Default FALSE. If true, then all possible index flavors are returned rather than just the ones that are currently defined.}
}
\value{
List of active indices, or all possible valid indices if 'valid' is set to TRUE
}
\description{
Return a list of indices defined on the triple-store
}

