% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_functions.R
\name{listNameSpaces}
\alias{listNameSpaces}
\title{listNameSpaces}
\usage{
listNameSpaces(service, catalogid = "root", repositoryid = "testWithParsa")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repositoryid}{Id for repository of interest.}
}
\value{
An ag object that includes the list of namespaces used in the specified repository.
}
\description{
listNameSpaces
}
\examples{
\dontrun{
service = createService("localhost","user","password")
listNameSpaces(service,catalogid = "root",repositoryid = "testRepo")
}
}

