% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{listNameSpaces}
\alias{listNameSpaces}
\title{Return all namespaces in repository}
\usage{
listNameSpaces(service, catalogid = "root", repo = "test")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}
}
\value{
An ag object that includes the list of namespaces used in the specified repository.
}
\description{
Return the currently active namespaces for your user in this repository
}
\examples{
\dontrun{
service = createService("localhost","user","password")
listNameSpaces(service,catalogid = "root",repo = "testRepo")
}
}

