% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{listRepositories}
\alias{listRepositories}
\title{List all repositories in chosen catalog}
\usage{
listRepositories(service, catalogid = "root")
}
\arguments{
\item{service}{Service object containing service url, username, and password}

\item{catalogid}{id for catalog of interest}
}
\value{
ag object
}
\description{
List all repositories in chosen catalog
}
\examples{
\dontrun{
service = createService("localhost","user","password")
listCatalogs(service)
}
}

