% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{listScripts}
\alias{listScripts}
\title{Return a list of scripts}
\usage{
listScripts(repository)
}
\arguments{
\item{repository}{Object of type repository specifying server details and repository to work on.}
}
\value{
list of scripts that have been added to this repository
}
\description{
Return the list of scripts that have been added to this repository. These scripts can either be in Lisp or Javascript code.
}
\examples{
\dontrun{
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
listScripts(rep)
}
}

