% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_funcs.R
\name{listScripts}
\alias{listScripts}
\title{Return a list of scripts}
\usage{
listScripts(service, catalogid = "root", repo = "")
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}
}
\value{
list of scripts that have been added to this repository
}
\description{
Return the list of scripts that have been added to this repository. These scripts can either be in Lisp or Javascript code.
}
\examples{
\dontrun{
service = createService("localhost","user","password")
listScripts(service,repo = "testRepo")
}
}

