% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_funcs.R
\name{optimizeIndices}
\alias{optimizeIndices}
\title{Optimize the indices}
\usage{
optimizeIndices(service, catalogid = "root", repo = "", level = 2,
  wait = FALSE, index = NULL)
}
\arguments{
\item{service}{Service object containing service url, username, and password.}

\item{catalogid}{Id for catalog of interest.}

\item{repo}{Id for repository of interest.}

\item{level}{Specifies how aggressive the optimization will be (more aggression, more optimization, but takes more time). Has to be a positive integer (Defaults to 2)}

\item{wait}{Default FALSE. If TRUE, then the service does not return until the optimization is complete}

\item{index}{The name of the index to operate on; this can be specified multiple times in a list}
}
\description{
Optimize the triple-store indices
}
\examples{
\dontrun{
service = createService('localhost','user','password')
optimizeIndices(service,repo = "testRepo",index = list("gspoi","i"),level = 2,wait = TRUE)
}
}

