% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_funcs.R
\name{Sessions}
\alias{Sessions}
\alias{closeSession}
\alias{commit}
\alias{getDescriptionSession}
\alias{isActive}
\alias{ping}
\alias{rollback}
\alias{startSession}
\title{Working with sessions}
\usage{
startSession(catalog, repo, autocommit = FALSE, lifetime = 100,
  loadInitFile = FALSE, script = NULL)

isActive(Session)

commit(Session)

rollback(Session)

closeSession(Session)

getDescriptionSession(Session)

ping(Session)
}
\arguments{
\item{catalog}{Object of type catalog specifying server details and catalog to work in}

\item{repo}{String that specifies what repository to start session on}

\item{autocommit}{Defaults to FALSE. Specify whether or not the session should use transactions.}

\item{lifetime}{The number of seconds the session can be idle before being shutdown and reclaimed.}

\item{loadInitFile}{If true, then the initfile will be loaded when the session starts.}

\item{script}{The name of a script file to load; may be specified multiple times.}

\item{Session}{Session object created by the startSession function}
}
\value{
A new repository object that points to the newly created session, belonging to current user. This session object can be substituted
for any function in which you were using the repository object.
}
\description{
Working with sessions
}
\examples{
\dontrun{

#Starting a session
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
session = startSession(rep,lifetime = 1000,script = "path/to/script")
}

\dontrun{

#Committing transactions to the repository
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
session = startSession(rep,lifetime = 1000)
commit(session)
}

\dontrun{

#Other useful functions
service = service("localhost","user","password")
rep = repository(catalog(service,"root"),"test")
session = startSession(rep,lifetime = 100)
getDescriptionSession(session)
isActive(session)
pring(session)
closeSession(session)
}
}

