% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createService.R
\name{summary.service}
\alias{summary.service}
\title{Summary of service object}
\usage{
\method{summary}{service}(x, ...)
}
\arguments{
\item{x}{Object of type service}

\item{...}{}
}
\value{
Print statement showing all elements of service
}
\description{
Summary of service object
}
\examples{
\dontrun{
service = createService("localhost","user","password")
summary(service)
}
}

