% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createService.R
\name{testConnection}
\alias{testConnection}
\title{testConnection}
\usage{
testConnection(s)
}
\arguments{
\item{s}{The service object}
}
\value{
Print message stating whether or not the test was succesful
}
\description{
testConnection
}
\examples{
\dontrun{
service = createService("localhost","user","password")
testConnection(service)
}
}

