% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createService.R
\name{testConnection.service}
\alias{testConnection.service}
\title{testConnection}
\usage{
\method{testConnection}{service}(x, ...)
}
\arguments{
\item{x}{The service object}

\item{...}{}
}
\value{
Print message stating whether or not the test was succesful
}
\description{
testConnection
}
\examples{
\dontrun{
service = createService("localhost","user","password")
testConnection(service)
}
}

